% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.dvmdostem.R
\name{adjust.runmask.dvmdostem}
\alias{adjust.runmask.dvmdostem}
\title{Adjust runmask for dvmdostem.}
\usage{
adjust.runmask.dvmdostem(siteDataPath, rundir, pixel_X, pixel_Y)
}
\arguments{
\item{siteDataPath}{path to expected/default runmask (one of dvmdostem standard input files)}

\item{rundir}{path to the location for this run (local run directory)}

\item{pixel_X}{the X coordinate of the pixel to turn on (1 based!)}

\item{pixel_Y}{the Y coordinate of the pixel to turn on (1 based!)}
}
\description{
Adjust the runmask for dvmdostem. This is necessary if you are
using a mutisite dvmdostem dataset (more than one grid cell/pixel)
and you are not forcing the community (cmt or vegetation) type. In 
other words you are using a vegetation map to determine the pixel's cmt
type. In this case you must make sure that for the site and PFTs
you have selected, the underlying veg map classifies the site as the
same community type of the PFT you have chosen to run.
}
\author{
Tobey Carman
}
