% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.dvmdostem.R
\name{convert.samples.dvmdostem}
\alias{convert.samples.dvmdostem}
\title{Convert samples for dvmdostem}
\usage{
convert.samples.dvmdostem(trait_values)
}
\arguments{
\item{trait_values}{a matrix or dataframe of samples from the trait distribution}
}
\value{
matrix or dataframe with values transformed
}
\description{
convert parameters, do unit conversions and update parameter names from PEcAn database default
to units/names within dvmdostem
}
\details{
Performs model specific unit conversions on a a list of trait values,
such as those provided to write.config
}
\author{
Shawn Serbin, Tobey Carman
}
