% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.dvmdostem.R
\name{enforce.runmask.cmt.vegmap.harmony}
\alias{enforce.runmask.cmt.vegmap.harmony}
\title{Make sure that selected run mask pixel, veg map pixel value and CMT type are all copasetic. The
function calls stop() if there is anything inconsistent, for example more tha one pixel is 
enabled in the run mask, or the enabled pixel's vegetation type does not match the 
vegetation/community type of the chosen PFTs.}
\usage{
enforce.runmask.cmt.vegmap.harmony(siteDataPath, rundir, cmtnum)
}
\arguments{
\item{siteDataPath}{is the path to the folder where we expect to find the dvmdostem input data files.}

\item{rundir}{is the path to the local running directory where customized files (config, parameters, 
runmask etc) are copied to.}

\item{cmtnum}{is the community type (vegetation type) that should be used for the run. Based on the
chosen PFT, and required to look up the correct parameters in the parameter files.}
}
\value{
none
}
\description{
Make sure that selected run mask pixel, veg map pixel value and CMT type are all copasetic. The
function calls stop() if there is anything inconsistent, for example more tha one pixel is 
enabled in the run mask, or the enabled pixel's vegetation type does not match the 
vegetation/community type of the chosen PFTs.
}
\author{
Tobey Carman
}
