% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.dvmdostem.R
\name{write.config.dvmdostem}
\alias{write.config.dvmdostem}
\title{Write dvmdostem model configuration files}
\usage{
write.config.dvmdostem(defaults = NULL, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
none
}
\description{
Writes a dvmdostem PEcAn config file.
}
\details{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Tobey Carman, Shawn Serbin
}
