% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.dvmdostem.R
\name{write.data2pecan.file}
\alias{write.data2pecan.file}
\title{Write data into PEcAn shaped output file.}
\usage{
write.data2pecan.file(y_starts, outdir, pecan_requested_vars,
  monthly_dvmdostem_outputs, yearly_dvmdostem_outputs, px_Y, px_X)
}
\arguments{
\item{y_starts}{a list of years, i.e.: 1901, 1902, 1903, etc.}

\item{outdir}{a path to the location where were we will look for dvmdostem outputs and write PEcAn outputs.}

\item{pecan_requested_vars}{comma separated string listing the variables to process (PEcAn names).}

\item{monthly_dvmdostem_outputs}{list of files available from dvmdostem at monthly resolution.}

\item{yearly_dvmdostem_outputs}{list of files available from dvmdostem at yearly resolution.}

\item{px_Y}{the pixel offset, Y (latitude) dimension.}

\item{px_X}{the pixel offset, X (longitude) dimension.}
}
\description{
Write data into PEcAn shaped output file.
}
\author{
Tobey Carman
}
