% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.dvmdostem.R
\name{model2netcdf.dvmdostem}
\alias{model2netcdf.dvmdostem}
\title{Code to convert dvmdostem netcdf output into into CF standard}
\usage{
model2netcdf.dvmdostem(outdir, runstart, runend, pecan_requested_vars)
}
\arguments{
\item{outdir}{Location of dvmdostem model output}

\item{runstart}{??}

\item{runend}{??}

\item{pecan_requested_vars}{a space separated string with names of the PEcAn variables to output.}
}
\description{
Code to convert dvmdostem netcdf output into into CF standard
}
\examples{
\dontrun{
# example code here?
}

}
\author{
Tobey Carman, Shawn Serbin
}
