% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.dvmdostem.R
\name{requested_vars_string2list}
\alias{requested_vars_string2list}
\title{Requested variables string to list conversion.}
\usage{
requested_vars_string2list(req_v_str, outspec_path)
}
\arguments{
\item{req_v_str}{A string, (comma or space separated?) of variables}

\item{outspec_path}{The path to an outspec file}
}
\value{
a list of the requested variables
}
\description{
Look up the "depends_on" in the output variable mapping, 
accumulate a list of dvmdostem variables to turn on to support 
the requested variables in the pecan.xml tag
}
\author{
Tobey Carman
}
