% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecan2dvmdostem_variable_mapping.R
\docType{data}
\name{vmap_reverse}
\alias{vmap_reverse}
\title{Build a mapping from dvmdostem names to PEcAn names, units, etc.
The temunits should be (are) looked up from the dvmdostem output 
file's units attributes.}
\format{
An object of class \code{list} of length 19.
}
\usage{
vmap_reverse
}
\description{
This data structure allows us to keep track of PEcAn output variables
that might depend on more than one DVMDOSTEM files.
}
\keyword{datasets}
